realization of task java oop advanced. Implemented methods: randomFigure.(3)

+++ src/main/java/core/basesyntax/Circle.java
@@ -9,8 +9,8 @@ public class Circle extends Figure {
     }
 
     @Override
-    public int getArea() {
-        return (int) Math.PI * radius * radius;
+    public double getArea() {
+        return Math.PI * radius * radius;
     }
 
     @Override
